% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_exploration.R
\name{timeSeriesExploration}
\alias{timeSeriesExploration}
\title{timeSeriesExploration
A plugin to make time series exploration easier}
\usage{
timeSeriesExploration()
}
\description{
timeSeriesExploration
A plugin to make time series exploration easier
}
